fix(repo): fix package generator
